Feat: Initial import from my `.tmux.conf`.
This code adds support for keybindings that are similar to `i3wm` in `tmux`.
It's not *exactly* the same as the keybindings in `i3wm`, since there are
some slight inconsistencies between the bindings used by `i3wm` and `vim`.
Also, the tiling models of `i3wm` and `tmux` are somewhat different. However,
I think this now gives a very good imitation of the `i3wm` work flow model.