Feat: Stricter adherence to layouts on `tmux` v2.x.
`tmux` v2.x has a "hook" feature, which lets us reapply a layout every time
a pane is created or destroyed, even if this is not done via `tilish`
keyboard shortcuts (e.g. using `C-d` in a terminal or `C-b "` in `tmux`).

This feature does not appear to be supported on `tmux` v1.x, so for now
you'll have to use `M-r` manually more often if using those versions.