Fix: Bug when closing last pane with M-S-q.
Previously, the non-legacy binding of M-S-q ran `kill-pane` followed
by two `select-layout` commands. However, this apparently causes the
newest (v3.1) version of `tmux` to freeze, and has the side-effect
that it reapplies layouts in other workspaces when closing the last
pane in a given workspace. This commit should fix that behavior.