Fix moving panes to workspaces in prefix mode on strict POSIX shells
The "expr substr" syntax used to return characters from a string is not defined
in the POSIX sh specication. Attempting to use that syntax on a Macbook led to
the bindings for moving panes between windows in prefix mode to fail to be
bound.

This patch addresses this by using a POSIX-compliant replacement for returning a
character from a string by index using the cut command.