preserve pane arrangement on refresh
A fresh window is configured with an undefined layout, and `select-layout` is
basically a noop. However, once a preset layout is applied then `select-layout`
will re-apply the last layout.

This commit allows the user to continue modifying the pane arrangement after a
preset layout has been applied by removing the automatic `select-layout` from
the hook.